+++ src/main/java/com/brandonitaly/locatorcolors/client/LocatorColorsClient.java
@@ -14,6 +14,11 @@ public class LocatorColorsClient /*? if fabric {*/ implements ClientModInitializ
     //? if fabric {
     @Override
     public void onInitializeClient() {
+        //? if >=26.1 {
+        net.fabricmc.fabric.api.client.keymapping.v1.KeyMappingHelper.registerKeyMapping(LocatorColorsKeyBindings.INSPECT_LOCATOR);
+        //?} else {
+        /*net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper.registerKeyBinding(LocatorColorsKeyBindings.INSPECT_LOCATOR);
+        *///?}
     }
     //?}
 
@@ -22,6 +27,9 @@ public class LocatorColorsClient /*? if fabric {*/ implements ClientModInitializ
         modContainer.registerExtensionPoint(IConfigScreenFactory.class, 
             (container, parent) -> new LocatorColorsConfigScreen(parent)
         );
+        modBus.addListener((net.neoforged.neoforge.client.event.RegisterKeyMappingsEvent event) -> {
+            event.register(LocatorColorsKeyBindings.INSPECT_LOCATOR);
+        });
     }
     *///?}
 }